+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - full stitches & ref compatibility (forwardRef, className forward, subcomponents selectors, displayName, toString) for:
   - `Input`
   - `HeaderIconAction`
+### Fixed
+- `Input` stale `onKeyDown` and `onSuggestionMatch` may be called
 
 ## [0.32.1] - 2025-12-19
 ### Fixed